Safari-611.1.10/LayoutTests/js/flatten-dictionary-structure-from-which-all-properties-were-deleted-expected.txt

Tests that deleting all properties from an object and then flattening it doesn't cause inconsistencies.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS p.f is 42
PASS successfullyParsed is true

TEST COMPLETE
